[datetime offset][s]: when applying datetime filter, consider local timezone offset.
This assumes that date[time] values in data  is in UTC timezone. For example, user wants to get timeseries data with hourly granularity from 1st of January 2020. Depending on user's local timezone, selected filter will be converted into UTC time so if he/she is in GMT+6, the data starting from 31st Dec 2019 18:00:00 will be received.
To fix that issue, we should consider timezone offset of the user.